feat: Implement validating if author is sending requests for someone else's tour

+++ src/Modules/Tours/Explorer.Tours.Core/UseCases/Administration/TourService.cs
@@ -4,6 +4,7 @@ using Explorer.Tours.API.Dtos;
 using Explorer.Tours.API.Public.Administration;
 using Explorer.Tours.Core.Domain;
 using Explorer.Tours.Core.Domain.RepositoryInterfaces;
+using System.Net.Sockets;
 namespace Explorer.Tours.Core.UseCases.Administration;
 
 public class TourService : ITourService
@@ -23,9 +24,12 @@ public class TourService : ITourService
         return _mapper.Map<TourDto>(result);
     }
 
-    public void Delete(long id)
+    public void Delete(long id, long authorId)
     {
-       _tourRepository.Delete(id);
+        var tour = _tourRepository.Get(id);
+        if (tour.CreatorId != authorId)
+            throw new InvalidOperationException("Can't delete someone else's tour");
+        _tourRepository.Delete(id);
     }
 
     public List<TourDto> GetAll()
@@ -53,9 +57,11 @@ public class TourService : ITourService
         return new PagedResult<TourDto>(items, result.TotalCount);
     }
 
-    public TourDto Update(long id, TourDto tourDto)
+    public TourDto Update(long id, TourDto tourDto, long authorId)
     {
         var tour = _tourRepository.Get(id);
+        if (tour.CreatorId != authorId)
+            throw new InvalidOperationException("Can't update someone else's tour");
         tour.Update(tourDto.CreatorId, tourDto.Title, tourDto.Description, tourDto.Difficulty,
             tourDto.Tags, (TourStatus)tourDto.Status, tourDto.Price);
 
@@ -63,27 +69,33 @@ public class TourService : ITourService
         return _mapper.Map<TourDto>(result);
     }
 
-    public KeypointDto AddKeypoint(long tourId, KeypointDto keypointDto)
+    public KeypointDto AddKeypoint(long tourId, KeypointDto keypointDto, long authorId)
     {
         var tour = _tourRepository.Get(tourId);
+        if (tour.CreatorId != authorId)
+            throw new InvalidOperationException("Can't add keypoint to someone else's tour");
         var keypoint = tour.AddKeypoint(_mapper.Map<Keypoint>(keypointDto));
         _tourRepository.Update(tour);
 
         return _mapper.Map<KeypointDto>(keypoint);
     }
 
-    public KeypointDto UpdateKeypoint(long tourId, KeypointDto keypointDto)
+    public KeypointDto UpdateKeypoint(long tourId, KeypointDto keypointDto, long authorId)
     {
         var tour = _tourRepository.Get(tourId);
+        if (tour.CreatorId != authorId)
+            throw new InvalidOperationException("Can't update keypoint from someone else's tour");
         var keypoint = tour.UpdateKeypoint(_mapper.Map<Keypoint>(keypointDto));
         _tourRepository.Update(tour);
 
         return _mapper.Map<KeypointDto>(keypoint);
     }
 
-    public void DeleteKeypoint(long tourId, long keypointId)
+    public void DeleteKeypoint(long tourId, long keypointId, long authorId)
     {
         var tour = _tourRepository.Get(tourId);
+        if (tour.CreatorId != authorId)
+            throw new InvalidOperationException("Can't delete keypoint from someone else's tour");
         tour.DeleteKeypoint(keypointId);
         var result = _tourRepository.Update(tour);
     }
